Added An actual ElementBuffer class
This is to enable VAOs to have serveral ElementBuffers to the same
VertexBuffer. So everything can be bound at once. Right now calls aren't
too worried about optimization there and there are only two calls:
A plain one that draws all element buffers in the VAO and one that takes
a string and only draws Element Buffers matching that name.
This is implemented and demonstraded now on GLDemoState. Pressing space
toggles between wireframe and fully rendered (controlled through EBOs)